Factor some shader-related functionality into a separate crate that we can also use from build.rs.
Differential Revision: https://phabricator.services.mozilla.com/D15368
