+++ webrender/Cargo.toml
@@ -16,7 +16,10 @@ capture = ["webrender_api/serialize", "ron", "serde", "debug_renderer"]
 replay = ["webrender_api/deserialize", "ron", "serde"]
 debug_renderer = []
 pathfinder = ["pathfinder_font_renderer", "pathfinder_gfx_utils", "pathfinder_partitioner", "pathfinder_path_utils"]
-serialize_program = ["serde"]
+serialize_program = ["serde", "webrender_build/serialize_program"]
+
+[build-dependencies]
+webrender_build = { version = "0.0.1", path = "../webrender_build" }
 
 [dependencies]
 app_units = "0.7"
@@ -42,6 +45,7 @@ smallvec = "0.6"
 thread_profiler = "0.1.1"
 time = "0.1"
 webrender_api = { version = "0.58.0", path = "../webrender_api" }
+webrender_build = { version = "0.0.1", path = "../webrender_build" }
 ws = { optional = true, version = "0.7.3" }
 
 [dependencies.pathfinder_font_renderer]
